feat: add revision range support
With this change, the user can specify a revision range of
commits to check, "git rev-list-style", like `HEAD ^HEAD~3`
or `HEAD~~..HEAD`.
It also provides the option to require a Jira-style ticket
to be present within that range.
Optionally, a custom list of conventional commit tags to
check for can be provided as well.

Implements: PIPE-935
Implements: PIPE-965