feat: add Hopic template option for commit range ticket checking
This introduces `require-ticket` as an optional parameter to the
template.
If set, when the build is a pull request (that is, it has a source
and a target), the commit range will be checked and commisery will
return non-zero if the commit messages do not reference a Jira-style
ticket.

Addresses: PIPE-965