fix: rework original commit.py to new rules checker
The recent redesign of the commit module broke public API without
good reason. Since the redesign was squashed down to a single commit and
contained functionality we'd like to keep, we can't easily partially roll
it back.

Therefore, not-so-trivial functional changes have had to be made to
restore `commit.py` to a state resembling the previous one.

This includes:
 - setting class instantiation back to `parse_commit_message`, which
   returns a specialized ConventionalCommit object with a fallback to
   a regular CommitMessage object
 - removing the verification code in the ConventionalCommit initializer,
   which is now fully handled by the rules module
 - adding accessor properties to the commit module, used by the rules
   module
 - re-writing the rules module to use functionality offered by the
   commit module
 - relaxing the strict Conventional Commit regex, so the rules module
   can catch (slight) errors